Added Link to Report Page on Popup (Unfinished): Part of Issue #25
Added link to popup that takes you to the report page. It does not pass
along the selection yet. Styling changed because I had to use the
description part of the PopupTemplate to add field data and hyperlink
instead of the fieldInfos section which comes with theme styling to
highlight the data from the field name. Should figure out how to remedy
this later