Part of Issue #99: Opens Report In New Window (Bug with dotnum variable not updating on each click)
Full Report link opens the report page in a new window correctly.
However, dotnum variable doesn't update on each subsequent popup
"click." If you don't refresh the index.html page before clicking on
another crossing and navigating to the report page, then it will
continue to open tabs that generate a report for the first crossing you
clicked on. Not sure how to fix but working on it.