Fix Grid Cell Alignment and Fix Comments Overflow Issue
Rows that have a cell where the label overflows onto a second line have
minor styling issues with the border lines. Adjusted cell alignment so
no labels overflow onto a second line.

When using the "input" tags for the Comments section, long comments
would be cut off. Changed tags to "textarea" so that the size of the
cell would increase to accommodate long comments.